OC-445: Move CorrelationIds to a more appropriate package.
